requirements.txt
torch
torchvision
transformers
Pillow
fastapi
uvicorn